Add files via upload
Nowadays, telecom industry faces fierce competition in satisfying its customers. The role of churn prediction is not only to accurately predict churners but also to interpret customer churn behavior.

The objectives of this project are using machine learning to predict and to find the root causes of churn. 

The dataset used is from kaggle. (https://www.kaggle.com/jpacse/datasets-for-churn-telecom)